Split scanner, filter, and frontmatter into focused submodules
Pure structural refactor of the three largest hyalo-core modules
(4660 lines total) into focused submodules. No behavioral changes —
all public APIs re-exported from mod.rs.

- scanner.rs → scanner/{mod,fence,visitor,frontmatter,strip}.rs
- filter.rs → filter/{mod,parse,match_props,tasks,sort,fields}.rs
- frontmatter.rs → frontmatter/{mod,parse,types}.rs
